fixed flipping bug
instead of rotating the cards, I had them as flipped X and Y, which is NOT THE SAME as rotating 180 degrees. As such, the movement patterns didn't look like they were supposed to. Originally, I fixed the movement to look like the cards. After figuring this out, I switched the movement back to the correct way, and rotated the cards instead of flipping X and Y.